Assistant checkpoint: Disabled resize for date input field
Assistant generated file changes:
- client/src/styles/App.css: Remove resize functionality from date input

---

User prompt:

This is a screenshot of the date field in the "Task" panel details view of a task. Note that their is a place in the lower right corner where you can grab and extend the field to make it larger. But that grab & stretch handle is ontop of the calendar icon to open the date picker. So I can't easily click the date picker. But there is no reason for the date field to be expandable. So, for only that 1 field, please remove the ability to drag and make it larger.