Assistant checkpoint: Remove order handling from date updates
Assistant generated file changes:
- client/src/components/Task.jsx: Remove order handling from date update

---

User prompt:

We just stripped out the drag and drop features with the intent to start over. However when i set the date on a task I get the attached error. It's understandable that it can't reach anything related to the drage and drop (like the order). But I don't know why setting the date would even be trying to do anything with task order. Changing the date doesn't change the order, you know? Can you tell why? and/or just remove that call to fix the error for now since we're disabling drag n drop.